# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VenueFlow API - Hospitality venue management SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List guests
      parameters:
        - name: venueId
          in: query
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: vipOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Create guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestBody"
      responses:
        "201":
          description: Created guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get guest detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    put:
      operationId: updateGuest
      tags: [guests]
      summary: Update guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuestBody"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Time Clock ────────────────────────────────────────────────────────────────
  /time-clock/in:
    post:
      operationId: clockIn
      tags: [timeClock]
      summary: Clock in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInBody"
      responses:
        "201":
          description: Clocked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeClockEntry"
...
  /floor-sections:
    get:
      operationId: listFloorSections
      tags: [floorPlan]
      summary: List floor sections
      parameters:
        - name: venueId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FloorSection"
    post:
      operationId: createFloorSection
      tags: [floorPlan]
      summary: Create floor section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFloorSectionBody"
      responses:
        "201":
          description: Created section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FloorSection"

  # ── Tables ────────────────────────────────────────────────────────────────────
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List tables
      parameters:
        - name: venueId
          in: query
          required: true
          schema:
            type: string
        - name: sectionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    put:
      operationId: updateTable
      tags: [tables]
      summary: Update table status or position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableBody"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Tip Pool ──────────────────────────────────────────────────────────────────
...
  /analytics/dashboard:
    get:
      operationId: getDashboardAnalytics
      tags: [analytics]
      summary: Dashboard summary stats
      parameters:
        - name: venueId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /analytics/labor:
    get:
      operationId: getLaborAnalytics
      tags: [analytics]
      summary: Labor hours and cost by date range
      parameters:
        - name: venueId
          in: query
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Labor analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LaborAnalyticsEntry"
  /analytics/employees:
    get:
      operationId: getEmployeeAnalytics
      tags: [analytics]
      summary: Per-employee hours, pay, attendance
      parameters:
        - name: venueId
          in: query
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeAnalyticsEntry"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Guest:
      type: object
      properties:
        id:
          type: string
        venueId:
          type: string
        fullName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        birthday:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        vipLevel:
          type: integer
          minimum: 0
          maximum: 3
        visitCount:
          type: integer
        lastVisit:
          type: string
          nullable: true
        totalSpent:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, venueId, fullName, tags, vipLevel, visitCount, totalSpent, createdAt]
    CreateGuestBody:
      type: object
      properties:
        venueId:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        birthday:
          type: string
        tags:
          type: array
          items:
            type: string
        vipLevel:
          type: integer
        notes:
          type: string
      required: [venueId, fullName]
    UpdateGuestBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        birthday:
          type: string
        tags:
          type: array
          items:
            type: string
        vipLevel:
          type: integer
        notes:
          type: string
        visitCount:
          type: integer
        totalSpent:
          type: number
    TimeClockEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
          nullable: true
        venueId:
          type: string
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: string
          format: date-time
          nullable: true
        totalHours:
          type: number
          nullable: true
        breakMinutes:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, completed]
        notes:
          type: string
          nullable: true
      required: [id, userId, venueId, clockIn, status]
    ClockInBody:
      type: object
      properties:
        userId:
          type: string
        venueId:
          type: string
        notes:
          type: string
      required: [userId, venueId]
...
    FloorSection:
      type: object
      properties:
        id:
          type: string
        venueId:
          type: string
        name:
          type: string
        capacity:
          type: integer
        color:
          type: string
      required: [id, venueId, name, capacity, color]
    CreateFloorSectionBody:
      type: object
      properties:
        venueId:
          type: string
        name:
          type: string
        capacity:
          type: integer
        color:
          type: string
      required: [venueId, name, capacity, color]
    Table:
      type: object
      properties:
        id:
          type: string
        venueId:
          type: string
        sectionId:
          type: string
        sectionName:
          type: string
          nullable: true
        label:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved, cleaning]
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
      required: [id, venueId, sectionId, label, capacity, status, x, y, width, height]
    CreateTableBody:
      type: object
      properties:
        venueId:
          type: string
        sectionId:
          type: string
        label:
          type: string
        capacity:
          type: integer
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
      required: [venueId, sectionId, label, capacity, x, y, width, height]
    UpdateTableBody:
      type: object
      properties:
        label:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved, cleaning]
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
...
    DashboardStats:
      type: object
      properties:
        activeStaffCount:
          type: integer
        shiftsToday:
          type: integer
        openShifts:
          type: integer
        laborPct:
          type: number
        waitlistCount:
          type: integer
        reservationsToday:
          type: integer
        clockedInNow:
          type: integer
        pendingTimeOff:
          type: integer
        pendingShiftRequests:
          type: integer
      required: [activeStaffCount, shiftsToday, openShifts, laborPct, waitlistCount, reservationsToday, clockedInNow, pendingTimeOff, pendingShiftRequests]
    LaborAnalyticsEntry:
      type: object
      properties:
        date:
          type: string
          format: date
        totalHours:
          type: number
        totalCost:
          type: number
        scheduledShifts:
          type: integer
      required: [date, totalHours, totalCost, scheduledShifts]
    EmployeeAnalyticsEntry:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        totalHours:
          type: number
        regularHours:
          type: number
        overtimeHours:
          type: number
        totalPay:
          type: number
        shiftsWorked:
          type: integer
      required: [userId, userName, totalHours, regularHours, overtimeHours, totalPay, shiftsWorked]