# lib/api-zod/src/generated/types/createGuestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VenueFlow API - Hospitality venue management SaaS
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGuestBody {
  venueId: string;
  fullName: string;
  email?: string;
  phone?: string;
  birthday?: string;
  tags?: string[];
  vipLevel?: number;
  notes?: string;
}