# lib/api-zod/src/generated/types/updateTableBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VenueFlow API - Hospitality venue management SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTableBodyStatus } from "./updateTableBodyStatus";

export interface UpdateTableBody {
  label?: string;
  capacity?: number;
  status?: UpdateTableBodyStatus;
  x?: number;
  y?: number;
  width?: number;
  height?: number;
}